Catch exception when fetching permissions
As this might cause problems when database structure has changed in
mapping, but users is attempting to execute doctrine:migrations commands
they get hit with exceptions and thus makes the commands unusable until
they change the vendor code. It's better in this situation to just catch
any exception and then logging it, hoping that users check their logs.